requirements.txt
requests~=2.32.3
aligo~=6.2.4
python-dotenv~=1.0.1